Add files via upload
Yeah so this got to ~80% accuracy which I think is fine, it saves every 5 epochs so there's a good chance there was one with higher accuracy but I picked that one to test and it was good enough. I don't think this is 100% stable because some of the later epochs get pretty bad but that's not really a major concern of mine although it really should be fixed.