Add files via upload
So I'm working on getting the tuning curve working when the input pattern has a phase offset, that is to say that the lines are horizontal or vertical but they aren't always in the same place. I added 20 inhibitory neurons and its at like 57% accuracy which isn't good but is pretty interesting. I need to check if that's better than if they aren't there, I think it is though.